+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- feat: start/stop integrations when calling optIn() or optOut() ([#329](https://github.com/PostHog/posthog-ios/pull/329))
+
 ## 3.22.1 - 2025-04-10
 
 - no user facing changes